+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralworks/integrations/powah/PowahEnergyStorageWrapper.kt
@@ -2,6 +2,7 @@ package site.siredvin.peripheralworks.integrations.powah
 
 import owmii.powah.lib.block.AbstractEnergyStorage
 import site.siredvin.broccolium.modules.storage.energy.AgnosticEnergyStack
+import site.siredvin.broccolium.modules.storage.energy.EnergyUnit
 import site.siredvin.broccolium.modules.storage.energy.api.AgnosticEnergyStorage
 import site.siredvin.peripheralworks.xplat.ModPlatform
 import java.util.function.Predicate
@@ -9,12 +10,19 @@ import java.util.function.Predicate
 class PowahEnergyStorageWrapper(private val storage: AbstractEnergyStorage<*, *>) : AgnosticEnergyStorage {
     override val capacity: Long
         get() = storage.energy.capacity
+    override val canExtract: Boolean
+        get() = storage.canExtractEnergy(null)
     override val energy: AgnosticEnergyStack
         get() = AgnosticEnergyStack(ModPlatform.commonEnergy, storage.energy.stored)
 
     override fun setChanged() {
         storage.setChanged()
     }
+
+    override val canReceive: Boolean
+        get() = storage.canReceiveEnergy(null)
+    override val unit: EnergyUnit
+        get() = ModPlatform.commonEnergy
 
     override fun storeEnergy(stack: AgnosticEnergyStack): AgnosticEnergyStack {
         if (!stack.`is`(ModPlatform.commonEnergy)) return stack
@@ -24,9 +32,9 @@ class PowahEnergyStorageWrapper(private val storage: AbstractEnergyStorage<*, *>
     }
 
     override fun takeEnergy(predicate: Predicate<AgnosticEnergyStack>, limit: Long): AgnosticEnergyStack {
-        if (!predicate.test(energy)) return AgnosticEnergyStack.EMPTY
+        if (!predicate.test(energy)) return AgnosticEnergyStack(ModPlatform.commonEnergy, 0)
         val extractedEnergy = storage.extractEnergy(limit, false, null)
-        if (extractedEnergy == 0L) return AgnosticEnergyStack.EMPTY
+        if (extractedEnergy == 0L) return AgnosticEnergyStack(ModPlatform.commonEnergy, 0)
         return AgnosticEnergyStack(ModPlatform.commonEnergy, extractedEnergy)
     }
 }