+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/ars_nouveau/Configuration.kt
@@ -8,12 +8,18 @@ object Configuration : IForgeConfigHandler {
         get() = "ars_nouveau"
 
     private var enableCasterTomePocketUpgradeConfig: ForgeConfigSpec.BooleanValue? = null
+    private var enableSourceStorageConfig: ForgeConfigSpec.BooleanValue? = null
 
     val enableCasterTomePocketUpgrade: Boolean
         get() = enableCasterTomePocketUpgradeConfig?.get() != false
 
+    val enableSourceStorage: Boolean
+        get() = enableSourceStorageConfig?.get() != false
+
     override fun addToConfig(builder: ForgeConfigSpec.Builder) {
         enableCasterTomePocketUpgradeConfig = builder.comment("Enables usage of caster tome as pocket computer upgrade")
             .define("enableCasterTomePocketUpgrade", true)
+        enableSourceStorageConfig = builder.comment("Enabled source storage")
+            .define("enableSourceStorage", true)
     }
 }