+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/theurgy/Integration.kt
@@ -1,6 +1,6 @@
-package site.siredvin.peripheralworks.integrations.embers
+package site.siredvin.peripheralworks.integrations.theurgy
 
-import com.rekindled.embers.api.capabilities.EmbersCapabilities
+import com.klikli_dev.theurgy.registry.CapabilityRegistry
 import site.siredvin.broccolium.modules.storage.energy.AgnosticEnergyStorageLookup
 import site.siredvin.broccolium.modules.storage.energy.api.AgnosticEnergyStorageExtractor
 import site.siredvin.peripheralworks.common.configuration.PeripheralWorksConfig
@@ -9,13 +9,13 @@ class Integration : Runnable {
 
     override fun run() {
         PeripheralWorksConfig.registerIntegrationConfiguration(Configuration)
-        if (Configuration.enableEmberStorage) {
+        if (Configuration.enableMercuryFluxStorage) {
             AgnosticEnergyStorageLookup.addEnergyStorageExtractor(
                 AgnosticEnergyStorageExtractor { level, blockPos, blockEntity ->
                     if (blockEntity == null) return@AgnosticEnergyStorageExtractor null
-                    val capability = blockEntity.getCapability(EmbersCapabilities.EMBER_CAPABILITY)
-                    if (capability.isPresent) {
-                        return@AgnosticEnergyStorageExtractor AgnosticEmberStorage(capability.resolve().get())
+                    val mfCapability = blockEntity.getCapability(CapabilityRegistry.MERCURY_FLUX)
+                    if (mfCapability.isPresent) {
+                        return@AgnosticEnergyStorageExtractor AgnosticMercuryFluxStorage(mfCapability.resolve().get())
                     }
                     return@AgnosticEnergyStorageExtractor null
                 },